requirements.txt
streamlit
pypdf
Pillow